fix(GiniHealthSDK): Move extension files to Extensions folder
