+++ HealthSDK/GiniHealthSDK/Sources/GiniHealthSDK/Core/Extensions/String.swift
@@ -12,3 +12,15 @@ extension String {
         return UIColor(hex: String.rgbaHexFrom(rgbHex: self))
     }
 }
+
+public extension String {
+    var numberValue: NSNumber? {
+        let formatter = NumberFormatter()
+        formatter.numberStyle = .decimal
+        return formatter.number(from: self)
+    }
+    
+    static func rgbaHexFrom(rgbHex: String) -> String {
+       return "#\(rgbHex)FF"
+   }
+}
